backend/requirements.txt
fastapi
uvicorn 
pydantic 
sentence-transformers
scikit-learn 
numpy 
pandas
pytesseract
Pillow
pdfplumber